% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{lr}
\alias{lr}
\title{Fitting Linear Regression Models}
\usage{
lr(Y, X, CI = TRUE, CI.level = 0.95, Rcpp = FALSE)
}
\arguments{
\item{Y}{a numeric vector, treated as outcome.}

\item{X}{a numerical matrix, treated as predictors.}

\item{CI}{logical, "TRUE" by default.
If "TRUE", the results will output the confidence interval of betas.}

\item{CI.level}{a numeric number specifying the alpha-level when calculate CIs of betas (0.95 by default).
Only use when "CI.beta = TRUE"}

\item{Rcpp}{"FALSE" by default.
If "TRUE", the function will use cpp code to speed up the computation.}
}
\value{
A list contains statistics from the regression model
}
\description{
`lr()` serves the purpose of fitting linear regression models. It can be used for linear regression,
determination of confidence intervals for relevant parameters, computation of \eqn{R^2},
and facilitation of the F Test.
}
\examples{
 # fit a regression model
 lr(state.x77[, 4], state.x77[, c(1:3)], Rcpp = FALSE)

}
