% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LR_cpp}
\alias{LR_cpp}
\title{Multiplies two doubles}
\usage{
LR_cpp(Y, X)
}
\arguments{
\item{Y}{outcome vector}

\item{X}{predictors matrix}
}
\value{
Regression results
}
\description{
Multiplies two doubles
}
